# Conflicts:
#	src/main/antlr/in.handyman.raven/gmr/Raven.g4
#	src/main/java/in/handyman/raven/lib/IntegratedNoiseModelApiAction.java
#	src/main/java/in/handyman/raven/lib/MultipartUploadAction.java
#	src/main/java/in/handyman/raven/lib/model/NoiseModel/NoiseModelConsumerProcess.java
#	src/main/java/in/handyman/raven/lib/model/NoiseModel/NoiseModelResponse.java
#	src/main/java/in/handyman/raven/lib/model/noiseModel/NoiseModelDataItem.java
#	src/main/java/in/handyman/raven/lib/model/noiseModel/NoiseModelOutput.java
#	src/main/java/in/handyman/raven/lib/model/noiseModel/NoiseModelOutputEntity.java
#	src/main/resources/config.properties

+++ src/main/java/in/handyman/raven/lib/model/qrextraction/QrInputEntity.java
@@ -17,7 +17,7 @@ public class QrInputEntity implements CoproProcessor.Entity {
     private String filePath;
     private Integer groupId;
     private String originId;
-    private Integer paperNo;
+    private Long paperNo;
     private String fileId;
     private Long rootPipelineId;
     private Long tenantId;
